country_guessr/django_rest/requirements.txt
blessed==1.15.0
botocore==1.12.30
cached-property==1.5.1
cement==2.8.2
certifi==2018.10.15
chardet==3.0.4
colorama==0.3.9
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
django-filters==0.2.1
djangorestframework==3.9.0
djangorestframework-gis==0.13
docopt==0.6.2
docutils==0.14
idna==2.6
jmespath==0.9.3
jsonschema==2.6.0
pathspec==0.5.5
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==0.9.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.53.0
